scripts/requirements.txt
psycopg2-binary>=2.9.0
requests>=2.28.0
protobuf>=3.20.0
